+++ R/package_logo.R
@@ -16,6 +16,12 @@
 #' @param brackets \[`logical(1)`\]\cr
 #' Curly brackets around the package name?
 #'
+#' @param background
+#' A `ggplot` object, the background of the sticker.
+#'
+#' @param s_x,s_y,s_width,s_height,white_around_sticker
+#' Passed on to \code{\link[hexSticker]{sticker}}.
+#'
 #' @return
 #' A \code{ggplot} object.
 #'
@@ -26,7 +32,15 @@
 #' @examples
 #' print(package_logo("my_package", brackets = TRUE))
 
-package_logo <- function(package_name, brackets = FALSE) {
+package_logo <- function(
+    package_name, brackets = FALSE,
+    background = ggplot2::ggplot() + ggplot2::theme_void(),
+    s_x = 1,
+    s_y = 1,
+    s_width = 1,
+    s_height = 1,
+    white_around_sticker = FALSE
+  ) {
 
   ### input checks
   input_check_response(
@@ -36,6 +50,10 @@ package_logo <- function(package_name, brackets = FALSE) {
   input_check_response(
     check = checkmate::check_flag(brackets),
     var_name = "brackets"
+  )
+  input_check_response(
+    check = checkmate::check_class(background, "ggplot"),
+    var_name = "background"
   )
 
   ### define path
@@ -58,12 +76,11 @@ package_logo <- function(package_name, brackets = FALSE) {
     hexSticker::sticker(
 
       ### image
-      subplot = ggplot2::ggplot() +
-        ggplot2::theme_void(),
-      s_x = 1,
-      s_y = 1,
-      s_width = 2,
-      s_height = 2,
+      subplot = background,
+      s_x = s_x,
+      s_y = s_y,
+      s_width = s_width,
+      s_height = s_height,
 
       ### package name
       package = package_name,
@@ -77,13 +94,8 @@ package_logo <- function(package_name, brackets = FALSE) {
       h_size = 1.2,
       h_fill = "white",
       h_color = "black",
-      spotlight = TRUE,
-      l_x = 0.9,
-      l_y = 1.4,
-      l_width = 2,
-      l_height = 1,
-      l_alpha = 0.8,
-      white_around_sticker = FALSE,
+      spotlight = FALSE,
+      white_around_sticker = white_around_sticker,
 
       ### URL
       url = "",